Import updateLocal module
